+++ src/components/ui/label.tsx
@@ -1,5 +1,5 @@
 import * as React from "react"
-import { cn } from "#orck/lib/utils"
+import { cn } from "cn"
 
 function Label({ className, ...props }: React.ComponentProps<"label">) {
   return (